use an SVG which is yellow
When we switch from using the SVG inline to using it in an img, we can't simply set its color in CSS the way we were doing before.

Easy solution: edit the SVG to make the SVG itself yellow. It's not like we're changing the color of the logo all the time.

I left a copy of the black logo in the repository just in case.